// comonl-comonlwcl/src/app/modules/comonlapi/model/formRicercaDelega.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { StatoDelega } from './statoDelega';


export interface FormRicercaDelega { 
    codiceFiscaleDelegato?: string;
    cognomeDelegato?: string;
    nomeDelegato?: string;
    codiceFiscaleDelegante?: string;
    cognomeDelegante?: string;
    nomeDelegante?: string;
    codiceFiscaleImpresa?: string;
    denominazione?: string;
    statoDelegas?: Array<StatoDelega>;
}